sep file : is_concat_next()
